pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage"
version = "0.2.0"
description = "OpenEnv customer support triage benchmark with deterministic graders"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=1.30.0",
    "openenv-core[core]>=0.2.2",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "support_triage.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["support_triage", "support_triage.server"]
package-dir = { "support_triage" = ".", "support_triage.server" = "server" }
